Make v8::Locker imply v8::Isolate::Scope.
This is needed to fix a problem in multi-cage mode, where Locker's own constructor depends on the cage base poitners being set, which v8::Isolate::Scope normally does, but we can't create the scope outside the lock.

This is probably not the right solution, since it changes the semantics of v8::Locker in a way that requires embedder code changes. This probably can't be upstreamed as-is, so we should expect this commit to be reverted later.